+++ scripts/archiso-base-211216.sh
@@ -216,30 +216,30 @@ keymap_layout=jp
 
 [launcher]
 #icon=/usr/share/icons/hicolor/24x24/apps/gimp.png
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/gimp.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/gimp.png
 path=/usr/bin/gimp
 
 [launcher]
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/htop.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/htop.png
 path=/usr/bin/kitty htop
 
 [launcher]
 #icon=/usr/share/icons/hicolor/24x24/apps/org.inkscape.Inkscape.png
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/inkscape.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/inkscape.png
 path=/usr/bin/inkscape
 
 [launcher]
 #icon=/usr/share/icons/gnome/24x24/apps/utilities-terminal.png
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/kitty.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/kitty.png
 path=/usr/bin/kitty
 
 [launcher]
 #icon=/usr/share/icons/hicolor/24x24/apps/libreoffice-startcenter.png
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/libreoffice.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/libreoffice.png
 path=/usr/bin/libreoffice
 
 [launcher]
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/file-manager.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/file-manager.png
 path=/usr/bin/pcmanfm-qt
 
 [launcher]
@@ -248,11 +248,11 @@ path=/usr/bin/shotcut
 
 [launcher]
 #icon=/usr/share/icons/hicolor/24x24/apps/vivaldi.png
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/vivaldi.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/vivaldi.png
 path=/usr/bin/vivaldi-stable
 
 [launcher]
-icon=/usr/local/share/icons/Tela-circle/24x24/apps/vlc.png
+icon=/usr/share/icons/Tela-circle/24x24/apps/vlc.png
 path=/usr/bin/vlc
 
 [output]
@@ -342,11 +342,11 @@ xorg-xwayland
 (
   cd $(mktemp -d)
   git clone https://github.com/vinceliuice/Tela-circle-icon-theme.git .
-  mkdir -p "$OLDPWD/airootfs/usr/local/share/icons/Tela-circle/24x24/apps"
+  mkdir -p "$OLDPWD/airootfs/usr/share/icons/Tela-circle/24x24/apps"
   cat << /cat | xargs -I{} -P0 -n1 ffmpeg \
     -width 24 \
     -i src/scalable/apps/{}.svg \
-    "$OLDPWD/airootfs/usr/local/share/icons/Tela-circle/24x24/apps/{}.png"
+    "$OLDPWD/airootfs/usr/share/icons/Tela-circle/24x24/apps/{}.png"
 file-manager
 gimp
 htop